wiki-md: avoid duplicating wiki in filetype
The idea is to remove any occurrences of 'wiki.' in the current
filetype before adding 'wiki.' to it. But we have to match exactly.

The nested executes are a pain, and string escaping is even worse.

------------------------------------------------------------------------

This cropped up during testing where after

  $ vim ~/.wiki/index

Repeated `:edit` commands caused the filetype to become successively
longer (wiki. -> wiki.wiki. -> ...)! We would like to avoid this...